data persistence between scenes added
Created a startmenu scene with an input field and a
start button.

Moved the functionality found in the initial MainManager script to a newly created GameManager class/script.

Set up the MainManager script to handle data persistence between scenes by creating a static instance
"Instance" of it and, after a null check, assigning the current instance with "this" and calling DontDestroyOnLoad on the respective gameObject.

Created an empty holding the GameManager script in the main scene.

Removed the empty holding the MainManager script from the main scene and created it new in the startmenu scene.

Added a public string var playerNameString and a public int var topScore to the MainManager script.

Added a StartGame method to the MainManager script that loads the main scene and is called on click of the start button.

Inside the StartGame method, the playerNameString var is assigned the player name text which the user submitted through the input field.

Together with the MainManager, this playerNameString var is carried over to the main scene and accessed in the GameManager script with MainManager.Instance.playerNameString, to personalize the score text with the player name.

Using the likewise persistent topScore int var from the MainManager script in a conditional inside the GameManager script: if MainManager.Instance.topScore is smaller m_Points (the int var used by the GameManager scrpit to count up the score), assign the value of m_Points to it and use it in the best score text, else if it is greater m_Points, use the value of the topScore var in the best score text instead.
This allows for a beatable high score being active between replays (unless you manage to remove all pieces on the first go...).

That´s pretty much it.