data persistence between sessions added
Added a public string var topScorerNameString to the MainManager script.

Created a serializable class SaveData in the MainManager script, containing a public int var topScore and a public string var topScorerNameString, to handle the IO of said vars.

Added a SaveUserData method below the SaveData class,
in which:

- a new instance of SaveData data is created,

- the vars included in the SaveData class are assigned the values from their MainManager counterparts with data.topScore equals topScore and data.topScorerNameString equals topScorerNameString.

- a string var json is created and assigned the JSON version of data with JsonUtility.ToJson(data).

- a savefile.json is created from the json string at a given path with File.WriteAllText.

Added a LoadUserData method below the SaveUserData method,
in which:

- a string var path is created and assigned the path to the savefile.json.

- a conditional is added checking if the file exists with File.Exists(path).

inside the conditional:

- a string var json is created and assigned the content of the savefile with File.ReadAllText(path).

- an instance of SaveData data is created and assigned the converted content of json/ the savefile with JsonUtility.FromJson.

- the topScore var and the topScorerNameString var of the MainManager are assigned the respective values from the savefile with topScore equals data.topScore and topScorerNameString equals data.topScorerNameString.

This LoadUserData method is then called in the StartGame method of the MainManager script (startmenu scene/Start button OnClick - see previous commit), just before the main scene is being loaded.

The SaveUserData method is called in the GameManager script inside a newly created Exit method, which executes when the Q key is pressed and:

- exits playmode when running in the editor

- quits the application when running built/compiled.

In the GameOver method of the GameManager script, assigning the playerNameString to the topScorerNameString in the first part of the conditional (if m_Points (the current score) is greater MainManager.Instance.topScore (the top score) - see previous commit) with:
MainManger.Instance.topScorerNameString equals MainManager.Instance.playerNameString.

In the second part of the conditional (else if m_Points smaller MainManager.Instance.topScore), the topScorerNameString is used in the best score text instead of the playerNameString.

On playtesting/recording, it retained the player name and best score between scenes (pressing Space on game over) and properly loaded the top scorer name and top score from the savefile on quit (Exit method/pressing Q on game over) and restart.

That´s pretty much it.